[CMA-41] feat : home 수정중,  통계 버그픽스
