+++ app/src/main/java/com/climus/climeet/presentation/ui/main/global/searchprofile/SearchProfileViewModel.kt
@@ -1,4 +1,4 @@
-package com.climus.climeet.presentation.ui.main.home.search.viewpager
+package com.climus.climeet.presentation.ui.main.global.searchprofile
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
@@ -10,9 +10,7 @@ import com.climus.climeet.data.repository.MainRepository
 import com.climus.climeet.presentation.ui.intro.signup.climer.followcrag.FollowCragEvent
 import com.climus.climeet.presentation.ui.intro.signup.climer.model.FollowCrag
 import com.climus.climeet.presentation.ui.intro.signup.climer.toFollowCrag
-import com.climus.climeet.presentation.ui.main.home.search.SearchCragUiState
-import com.climus.climeet.presentation.ui.main.home.search.model.FollowClimber
-import com.climus.climeet.presentation.ui.main.home.search.toFollowClimber
+import com.climus.climeet.presentation.ui.main.global.searchprofile.model.FollowClimber
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
@@ -28,26 +26,27 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-
-data class SearchClimberUiState(
+data class SearchProfileUiState(
+    val routeList: List<BestRouteDetailInfoResponse> = emptyList(),
     val followingList: List<UserFollowSimpleResponse> = emptyList(),
+    val searchList: List<FollowCrag> = emptyList(),
     val searchClimberList: List<FollowClimber> = emptyList(),
     val progressState: Boolean = false,
     val emptyResultState: Boolean = false,
 )
 
-sealed class SearchClimberEvent{
-    data class ShowToastMessage(val msg: String): SearchCragEvent()
+sealed class SearchProfileEvent{
+    data class ShowToastMessage(val msg: String): SearchProfileEvent()
 }
 
 @HiltViewModel
-class SearchClimberViewModel @Inject constructor(private val repository: MainRepository): ViewModel() {
+class SearchProfileViewModel @Inject constructor(private val repository: MainRepository): ViewModel() {
 
-    private val _uiState = MutableStateFlow(SearchClimberUiState())
-    val uiState: StateFlow<SearchClimberUiState> = _uiState.asStateFlow()
+    private val _uiState = MutableStateFlow(SearchProfileUiState())
+    val uiState: StateFlow<SearchProfileUiState> = _uiState.asStateFlow()
 
-    private val _event = MutableSharedFlow<SearchClimberEvent>()
-    val event: SharedFlow<SearchClimberEvent> = _event.asSharedFlow()
+    private val _event = MutableSharedFlow<FollowCragEvent>()
+    val event: SharedFlow<FollowCragEvent> = _event.asSharedFlow()
 
     private var curJob: Job? = null
 
@@ -55,6 +54,26 @@ class SearchClimberViewModel @Inject constructor(private val repository: MainRep
 
     init {
         observeKeyword()
+    }
+
+    fun getRouteRankingOrderSelectionCount() {
+        viewModelScope.launch {
+            repository.findRouteRankingOrderSelectionCount().let {
+                when(it) {
+                    is BaseState.Success -> {
+                        _uiState.update { state ->
+                            state.copy(
+                                routeList = it.body
+                            )
+                        }
+                    }
+                    is BaseState.Error -> {
+                        it.msg // 서버 에러 메시지
+                        Log.d("API", it.msg)
+                    }
+                }
+            }
+        }
     }
 
     fun getClimberFollowing() {
@@ -77,11 +96,44 @@ class SearchClimberViewModel @Inject constructor(private val repository: MainRep
         }
     }
 
+    fun followUser(userId: Long) {
+        viewModelScope.launch {
+            repository.followUser(userId).let {
+                when(it) {
+                    is BaseState.Success -> {
+
+                    }
+                    is BaseState.Error -> {
+                        it.msg // 서버 에러 메시지
+                        Log.d("API", it.msg)
+                    }
+                }
+            }
+        }
+    }
+
+    fun unfollowUser(userId: Long) {
+        viewModelScope.launch {
+            repository.unfollowUser(userId).let {
+                when(it) {
+                    is BaseState.Success -> {
+
+                    }
+                    is BaseState.Error -> {
+                        it.msg // 서버 에러 메시지
+                        Log.d("API", it.msg)
+                    }
+                }
+            }
+        }
+    }
+
     private fun observeKeyword() {
         keyword.onEach {
             if (it.isBlank()) {
                 _uiState.update { state ->
                     state.copy(
+                        searchList = emptyList(),
                         searchClimberList = emptyList(),
                         emptyResultState = false
                     )
@@ -98,6 +150,40 @@ class SearchClimberViewModel @Inject constructor(private val repository: MainRep
 
                 curJob = viewModelScope.launch {
                     delay(500)
+                    repository.searchAvailableGym(it, 0, 15).let { result ->
+                        when (result) {
+                            is BaseState.Success -> {
+                                if (result.body.result.isNotEmpty()) {
+                                    _uiState.update { state ->
+                                        state.copy(
+                                            searchList = result.body.result.map { item ->
+                                                item.toFollowCrag(it)
+                                            },
+                                            progressState = false
+                                        )
+                                    }
+                                } else {
+                                    _uiState.update { state ->
+                                        state.copy(
+                                            searchList = emptyList(),
+                                            progressState = false,
+                                            emptyResultState = true
+                                        )
+                                    }
+                                }
+                            }
+
+                            is BaseState.Error -> {
+                                _uiState.update { state ->
+                                    state.copy(
+                                        progressState = false,
+                                        emptyResultState = true
+                                    )
+                                }
+
+                            }
+                        }
+                    }
 
                     repository.getClimberSearchingList(0, 15, it).let { result ->
                         when (result) {
@@ -141,8 +227,10 @@ class SearchClimberViewModel @Inject constructor(private val repository: MainRep
         keyword.value = ""
         _uiState.update { state ->
             state.copy(
+                searchList = emptyList(),
                 searchClimberList = emptyList()
             )
         }
     }
+
 }
