+++ app/src/main/java/com/climus/climeet/presentation/ui/main/global/searchprofile/recycler/FollowGymRVAdapter.kt
@@ -1,24 +1,21 @@
-package com.climus.climeet.presentation.ui.main.home.search.recycler
+package com.climus.climeet.presentation.ui.main.global.searchprofile.recycler
 
-import android.graphics.Color
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.climus.climeet.R
 import com.climus.climeet.data.model.response.UserHomeGymDetailResponse
 import com.climus.climeet.databinding.ItemFollowGymBinding
-import com.climus.climeet.presentation.ui.main.home.popularroutes.adapter.PopularRoutesAllRVadapter
 
 class FollowGymRVAdapter (private val followGymList: List<UserHomeGymDetailResponse>) : RecyclerView.Adapter<FollowGymRVAdapter.ViewHolder>() {
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FollowGymRVAdapter.ViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val binding: ItemFollowGymBinding = ItemFollowGymBinding.inflate(
             LayoutInflater.from(parent.context), parent, false)
         return ViewHolder(binding)
     }
 
-    override fun onBindViewHolder(holder: FollowGymRVAdapter.ViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         holder.bind(followGymList[position], position)
     }
