+++ app/src/main/java/com/climus/climeet/presentation/ui/main/global/searchprofile/recycler/RouteInfoRVAdapter.kt
@@ -1,27 +1,22 @@
-package com.climus.climeet.presentation.ui.main.home.search.recycler
+package com.climus.climeet.presentation.ui.main.global.searchprofile.recycler
 
-import android.graphics.Color
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
 import com.climus.climeet.R
 import com.climus.climeet.data.model.response.RouteSimpleInfo
-import com.climus.climeet.data.model.response.UserHomeGymDetailResponse
-import com.climus.climeet.databinding.ItemFollowGymBinding
 import com.climus.climeet.databinding.ItemPopularRoutesBinding
-import com.climus.climeet.presentation.ui.main.home.popularroutes.adapter.PopularRoutesAllRVadapter
 import com.climus.climeet.presentation.util.Constants
 
 class RouteInfoRVAdapter (private val routeInfos: List<RouteSimpleInfo>, private val gymName : String) : RecyclerView.Adapter<RouteInfoRVAdapter.ViewHolder>() {
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RouteInfoRVAdapter.ViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val binding: ItemPopularRoutesBinding = ItemPopularRoutesBinding.inflate(
             LayoutInflater.from(parent.context), parent, false)
         return ViewHolder(binding)
     }
 
-    override fun onBindViewHolder(holder: RouteInfoRVAdapter.ViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         holder.bind(routeInfos[position], position)
     }
