+++ app/src/main/java/com/climus/climeet/presentation/ui/main/global/searchprofile/viewpager/SearchClimberViewModel.kt
@@ -1,17 +1,13 @@
-package com.climus.climeet.presentation.ui.main.home.search.viewpager
+package com.climus.climeet.presentation.ui.main.global.searchprofile.viewpager
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.climus.climeet.data.model.BaseState
 import com.climus.climeet.data.model.response.UserFollowSimpleResponse
-import com.climus.climeet.data.model.response.UserHomeGymDetailResponse
 import com.climus.climeet.data.repository.MainRepository
-import com.climus.climeet.presentation.ui.intro.signup.climer.followcrag.FollowCragEvent
-import com.climus.climeet.presentation.ui.intro.signup.climer.model.FollowCrag
-import com.climus.climeet.presentation.ui.intro.signup.climer.toFollowCrag
-import com.climus.climeet.presentation.ui.main.home.search.model.FollowClimber
-import com.climus.climeet.presentation.ui.main.home.search.toFollowClimber
+import com.climus.climeet.presentation.ui.main.global.searchprofile.model.FollowClimber
+import com.climus.climeet.presentation.ui.main.global.searchprofile.toFollowClimber
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
@@ -27,26 +23,26 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-data class SearchCragUiState(
-    val followGymList: List<UserHomeGymDetailResponse> = emptyList(),
+
+data class SearchClimberUiState(
     val followingList: List<UserFollowSimpleResponse> = emptyList(),
-    val searchList: List<FollowCrag> = emptyList(),
+    val searchClimberList: List<FollowClimber> = emptyList(),
     val progressState: Boolean = false,
     val emptyResultState: Boolean = false,
 )
 
-sealed class SearchCragEvent{
+sealed class SearchClimberEvent{
     data class ShowToastMessage(val msg: String): SearchCragEvent()
 }
 
 @HiltViewModel
-class SearchCragViewModel @Inject constructor(private val repository: MainRepository): ViewModel() {
+class SearchClimberViewModel @Inject constructor(private val repository: MainRepository): ViewModel() {
 
-    private val _uiState = MutableStateFlow(SearchCragUiState())
-    val uiState: StateFlow<SearchCragUiState> = _uiState.asStateFlow()
+    private val _uiState = MutableStateFlow(SearchClimberUiState())
+    val uiState: StateFlow<SearchClimberUiState> = _uiState.asStateFlow()
 
-    private val _event = MutableSharedFlow<SearchCragEvent>()
-    val event: SharedFlow<SearchCragEvent> = _event.asSharedFlow()
+    private val _event = MutableSharedFlow<SearchClimberEvent>()
+    val event: SharedFlow<SearchClimberEvent> = _event.asSharedFlow()
 
     private var curJob: Job? = null
 
@@ -56,14 +52,14 @@ class SearchCragViewModel @Inject constructor(private val repository: MainReposi
         observeKeyword()
     }
 
-    fun getGymFollowing() {
+    fun getClimberFollowing() {
         viewModelScope.launch {
-            repository.getGymsFollowing().let {
+            repository.getClimberFollowing().let {
                 when(it) {
                     is BaseState.Success -> {
                         _uiState.update { state ->
                             state.copy(
-                                followGymList = it.body
+                                followingList = it.body
                             )
                         }
                     }
@@ -81,7 +77,7 @@ class SearchCragViewModel @Inject constructor(private val repository: MainReposi
             if (it.isBlank()) {
                 _uiState.update { state ->
                     state.copy(
-                        searchList = emptyList(),
+                        searchClimberList = emptyList(),
                         emptyResultState = false
                     )
                 }
@@ -97,14 +93,15 @@ class SearchCragViewModel @Inject constructor(private val repository: MainReposi
 
                 curJob = viewModelScope.launch {
                     delay(500)
-                    repository.searchAvailableGym(it, 0, 15).let { result ->
+
+                    repository.getClimberSearchingList(0, 15, it).let { result ->
                         when (result) {
                             is BaseState.Success -> {
                                 if (result.body.result.isNotEmpty()) {
                                     _uiState.update { state ->
                                         state.copy(
-                                            searchList = result.body.result.map { item ->
-                                                item.toFollowCrag(it)
+                                            searchClimberList = result.body.result.map { item ->
+                                                item.toFollowClimber(it)
                                             },
                                             progressState = false
                                         )
@@ -112,7 +109,7 @@ class SearchCragViewModel @Inject constructor(private val repository: MainReposi
                                 } else {
                                     _uiState.update { state ->
                                         state.copy(
-                                            searchList = emptyList(),
+                                            searchClimberList = emptyList(),
                                             progressState = false,
                                             emptyResultState = true
                                         )
@@ -127,7 +124,6 @@ class SearchCragViewModel @Inject constructor(private val repository: MainReposi
                                         emptyResultState = true
                                     )
                                 }
-
                             }
                         }
                     }
@@ -140,7 +136,7 @@ class SearchCragViewModel @Inject constructor(private val repository: MainReposi
         keyword.value = ""
         _uiState.update { state ->
             state.copy(
-                searchList = emptyList(),
+                searchClimberList = emptyList()
             )
         }
     }
