+++ app/src/main/java/com/climus/climeet/presentation/ui/main/home/recycler/following/FollowingRVAdapter.kt
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.climus.climeet.data.model.response.UserFollowSimpleResponse
 import com.climus.climeet.databinding.ItemFollowingBinding
-import com.climus.climeet.presentation.ui.main.home.search.model.FollowClimber
+import com.climus.climeet.presentation.ui.main.global.searchprofile.model.FollowClimber
 
 class FollowingRVAdapter(private val followingList: List<UserFollowSimpleResponse>) : RecyclerView.Adapter<FollowingRVAdapter.ViewHolder>(){
