+++ app/src/main/java/com/climus/climeet/presentation/ui/main/mypage/follow/viewpager/FollowerViewModel.kt
@@ -1,11 +1,9 @@
 package com.climus.climeet.presentation.ui.main.mypage.follow.viewpager
 
 import androidx.lifecycle.ViewModel
-import com.climus.climeet.data.model.response.BestClearClimberSimpleResponse
 import com.climus.climeet.data.model.response.UserFollowSimpleResponse
 import com.climus.climeet.data.repository.MainRepository
-import com.climus.climeet.presentation.ui.main.home.search.model.FollowClimber
-import com.climus.climeet.presentation.ui.main.home.viewpager.ranking.CompleteClimbingUiState
+import com.climus.climeet.presentation.ui.main.global.searchprofile.model.FollowClimber
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
